Refactor components into appropriate folders
