+++ src/electionguard_gui/web/components/admin/admin-home-component.js
