+++ src/electionguard_gui/web/components/admin/create-key-component.js
@@ -1,4 +1,4 @@
-import Spinner from "./spinner-component.js";
+import Spinner from "../shared/spinner-component.js";
 
 export default {
   components: {
