+++ src/electionguard_gui/web/components/admin/setup-election-component.js
