+++ src/electionguard_gui/web/components/guardian/guardian-home-component.js
