+++ src/electionguard_gui/web/components/shared/navbar-component.js
