+++ src/electionguard_gui/web/components/shared/select-mode-component.js
