+++ src/electionguard_gui/web/components/shared/spinner-component.js
